@@ -6,38 +6,46 @@
   "scripts": {
     "test": "echo \"Error: no test specified\" && exit 1"
   },
+  "babel": {
+    "presets": [
+      "es2015"
+    ]
+  },
   "dependencies": {},
   "devDependencies": {
     "babel-core": "^6.25.0",
     "babel-loader": "^7.0.0",
+    "babel-plugin-transform-runtime": "^6.23.0",
+    "babel-polyfill": "^6.26.0",
     "babel-preset-es2015": "^6.24.1",
     "babel-preset-es2015-riot": "^1.1.0",
     "browser-sync": "^2.18.8",
-    "del": "^2.2.2",
+    "del": "^3.0.0",
     "gulp": "^3.9.1",
-    "gulp-autoprefixer": "^3.1.1",
+    "gulp-autoprefixer": "^4.0.0",
     "gulp-changed": "^3.0.0",
     "gulp-concat": "^2.6.1",
-    "gulp-connect-php": "1.0.0",
-    "gulp-cssmin": "^0.1.7",
-    "gulp-imagemin": "^3.2.0",
+    "gulp-connect-php": "1.0.1",
+    "gulp-cssmin": "^0.2.0",
+    "gulp-imagemin": "^4.0.0",
     "gulp-plumber": "^1.1.0",
     "gulp-rename": "^1.2.2",
     "gulp-sass": "^3.1.0",
     "gulp-sftp": "^0.1.5",
     "gulp-sourcemaps": "^2.6.0",
-    "gulp-uglify": "^2.1.2",
+    "gulp-uglify": "^3.0.0",
     "gulp-util": "^3.0.8",
     "imagemin-pngquant": "^5.0.0",
     "jquery": "^3.2.1",
+    "node-sass": "^4.7.2",
     "npm-check-updates": "^2.11.0",
-    "riot": "^3.4.4",
+    "riot": "^3.7.4",
     "riot-route": "^3.1.1",
+    "riot-tag-loader": "^1.0.0",
     "riotcontrol": "0.0.3",
-    "riotjs-loader": "^4.0.0",
     "vinyl-ftp": "^0.6.0",
-    "webpack": "^2.6.1",
-    "webpack-stream": "^3.2.0"
+    "webpack": "^3.9.1",
+    "webpack-stream": "^4.0.0"
   },
   "author": "",
   "license": "ISC"
